+++ src/Features/PaymentGatewaySuggestions/DataSourcePoller.php
@@ -12,6 +12,14 @@ defined( 'ABSPATH' ) || exit;
  * This handles polling specs from JSON endpoints.
  */
 class DataSourcePoller {
+	/**
+	 * Name of data sources filter.
+	 */
+	const FILTER_NAME = 'woocommerce_admin_payment_gateway_suggestions_data_sources';
+
+	/**
+	 * Default data sources array.
+	 */
 	const DATA_SOURCES = array(
 		'https://woocommerce.com/wp-json/wccom/payment-methods/1.0/methods.json',
 	);
@@ -43,10 +51,10 @@ class DataSourcePoller {
 	 */
 	public static function read_specs_from_data_sources() {
 		$specs        = array();
-		$data_sources = apply_filters( 'woocommerce_admin_payment_gateway_suggestions_data_sources', self::DATA_SOURCES );
+		$data_sources = apply_filters( self::FILTER_NAME, self::DATA_SOURCES );
 
 		// Note that this merges the specs from the data sources based on the
-		// key - last one wins.
+		// id - last one wins.
 		foreach ( $data_sources as $url ) {
 			$specs_from_data_source = self::read_data_source( $url );
 			self::merge_specs( $specs_from_data_source, $specs, $url );
@@ -115,8 +123,8 @@ class DataSourcePoller {
 				continue;
 			}
 
-			$key           = $spec->key;
-			$specs[ $key ] = $spec;
+			$id           = $spec->id;
+			$specs[ $id ] = $spec;
 		}
 	}
 
@@ -132,9 +140,9 @@ class DataSourcePoller {
 		$logger         = self::get_logger();
 		$logger_context = array( 'source' => $url );
 
-		if ( ! isset( $spec->key ) ) {
+		if ( ! isset( $spec->id ) ) {
 			$logger->error(
-				'Spec is invalid because the key is missing in feed',
+				'Spec is invalid because the id is missing in feed',
 				$logger_context
 			);
 			// phpcs:ignore