+++ client/task-list/index.js
@@ -12,7 +12,9 @@ import {
 	PLUGINS_STORE_NAME,
 	SETTINGS_STORE_NAME,
 } from '@woocommerce/data';
+import { useExperiment } from '@woocommerce/explat';
 import { recordEvent } from '@woocommerce/tracks';
+import { ScrollTo } from '@woocommerce/components';
 
 /**
  * Internal dependencies
@@ -24,6 +26,7 @@ import { getCountryCode } from '../dashboard/utils';
 import TaskList from './task-list';
 import { DisplayOption } from '../header/activity-panel/display-options';
 import { TaskStep } from './task-step';
+import TaskListPlaceholder from './placeholder';
 
 const taskDashboardSelect = ( select ) => {
 	const { getProfileItems, getTasksStatus } = select( ONBOARDING_STORE_NAME );
@@ -43,6 +46,14 @@ const taskDashboardSelect = ( select ) => {
 	const countryCode = getCountryCode(
 		generalSettings.woocommerce_default_country
 	);
+	const {
+		woocommerce_store_address: storeAddress,
+		woocommerce_default_country: defaultCountry,
+		woocommerce_store_postcode: storePostCode,
+	} = generalSettings;
+	const hasCompleteAddress = Boolean(
+		storeAddress && defaultCountry && storePostCode
+	);
 
 	const activePlugins = getActivePlugins();
 	const installedPlugins = getInstalledPlugins();
@@ -52,6 +63,9 @@ const taskDashboardSelect = ( select ) => {
 		activePlugins,
 		countryCode,
 		dismissedTasks: getOption( 'woocommerce_task_list_dismissed_tasks' ),
+		remindMeLaterTasks: getOption(
+			'woocommerce_task_list_remind_me_later_tasks'
+		),
 		isExtendedTaskListComplete:
 			getOption( 'woocommerce_extended_task_list_complete' ) === 'yes',
 		isExtendedTaskListHidden:
@@ -65,6 +79,7 @@ const taskDashboardSelect = ( select ) => {
 		onboardingStatus,
 		profileItems,
 		trackedCompletedTasks,
+		hasCompleteAddress,
 	};
 };
 
@@ -82,12 +97,17 @@ const TaskDashboard = ( { userPreferences, query } ) => {
 		profileItems,
 		isSetupTaskListHidden,
 		dismissedTasks,
+		remindMeLaterTasks,
 		isTaskListComplete,
 		isExtendedTaskListHidden,
 		isExtendedTaskListComplete,
+		hasCompleteAddress,
 	} = useSelect( taskDashboardSelect );
 
 	const [ isCartModalOpen, setIsCartModalOpen ] = useState( false );
+	const [ isLoadingExperiment, experimentAssignment ] = useExperiment(
+		'woocommerce_tasklist_progression'
+	);
 
 	useEffect( () => {
 		document.body.classList.add( 'woocommerce-onboarding' );
@@ -161,6 +181,23 @@ const TaskDashboard = ( { userPreferences, query } ) => {
 		return currentTask;
 	};
 
+	const getExtendedTaskList = ( extensionTasks ) => {
+		return (
+			<TaskList
+				name="extended_task_list"
+				eventName="extended_tasklist"
+				collapsible
+				dismissedTasks={ dismissedTasks || [] }
+				remindMeLaterTasks={ remindMeLaterTasks || [] }
+				isComplete={ isExtendedTaskListComplete }
+				query={ query }
+				tasks={ extensionTasks }
+				title={ __( 'Things to do next', 'woocommerce-admin' ) }
+				trackedCompletedTasks={ trackedCompletedTasks || [] }
+			/>
+		);
+	};
+
 	const allTasks = getAllTasks( {
 		activePlugins,
 		countryCode,
@@ -173,6 +210,7 @@ const TaskDashboard = ( { userPreferences, query } ) => {
 		query,
 		toggleCartModal,
 		onTaskSelect,
+		hasCompleteAddress,
 	} );
 
 	const { extension, setup: setupTasks } = allTasks;
@@ -195,18 +233,47 @@ const TaskDashboard = ( { userPreferences, query } ) => {
 		);
 	}
 
+	const scrollToExtendedList =
+		window.location.hash.substr( 1 ) === 'extended_task_list';
+
 	return (
 		<>
-			{ setupTasks && ( ! isSetupTaskListHidden || task ) && (
-				<TaskList
-					dismissedTasks={ dismissedTasks || [] }
-					isComplete={ isTaskListComplete }
-					query={ query }
-					tasks={ setupTasks }
-					title={ __( 'Finish setup', 'woocommerce-admin' ) }
-					trackedCompletedTasks={ trackedCompletedTasks || [] }
-				/>
-			) }
+			{ setupTasks &&
+				( ! isSetupTaskListHidden || task ) &&
+				( isLoadingExperiment ? (
+					<TaskListPlaceholder />
+				) : (
+					<TaskList
+						name="task_list"
+						eventName="tasklist"
+						expandingItems={
+							experimentAssignment?.variationName === 'treatment'
+						}
+						dismissedTasks={ dismissedTasks || [] }
+						remindMeLaterTasks={ remindMeLaterTasks || [] }
+						isComplete={ isTaskListComplete }
+						query={ query }
+						tasks={ setupTasks }
+						title={ __(
+							'Get ready to start selling',
+							'woocommerce-admin'
+						) }
+						trackedCompletedTasks={ trackedCompletedTasks || [] }
+						onComplete={ () =>
+							updateOptions( {
+								woocommerce_default_homepage_layout:
+									'two_columns',
+							} )
+						}
+						onHide={ () =>
+							updateOptions( {
+								woocommerce_task_list_prompt_shown: true,
+								woocommerce_default_homepage_layout:
+									'two_columns',
+							} )
+						}
+					/>
+				) ) }
 			{ extensionTasks && (
 				<DisplayOption>
 					<MenuGroup
@@ -228,17 +295,15 @@ const TaskDashboard = ( { userPreferences, query } ) => {
 					</MenuGroup>
 				</DisplayOption>
 			) }
-			{ extensionTasks && ! isExtendedTaskListHidden && (
-				<TaskList
-					dismissedTasks={ dismissedTasks || [] }
-					isComplete={ isExtendedTaskListComplete }
-					name={ 'extended_task_list' }
-					query={ query }
-					tasks={ extensionTasks }
-					title={ __( 'Things to do next', 'woocommerce-admin' ) }
-					trackedCompletedTasks={ trackedCompletedTasks || [] }
-				/>
-			) }
+			{ extensionTasks &&
+				! isExtendedTaskListHidden &&
+				( scrollToExtendedList ? (
+					<ScrollTo offset="-20">
+						{ getExtendedTaskList( extensionTasks ) }
+					</ScrollTo>
+				) : (
+					getExtendedTaskList( extensionTasks )
+				) ) }
 			{ isCartModalOpen && (
 				<CartModal
 					onClose={ () => toggleCartModal() }