+++ client/task-list/tasks/payments/RemotePayments/components/RecommendedPaymentGatewayList/index.js
@@ -1,1 +1,0 @@
-export { RecommendedPaymentGatewayList } from './RecommendedPaymentGatewayList';