+++ packages/admin-e2e-tests/src/pages/OnboardingWizard.ts
@@ -6,7 +6,6 @@ import { Page } from 'puppeteer';
 /**
  * Internal dependencies
  */
-import { BenefitsSection } from '../sections/onboarding/BenefitsSection';
 import { BusinessSection } from '../sections/onboarding/BusinessSection';
 import { IndustrySection } from '../sections/onboarding/IndustrySection';
 import { ProductTypeSection } from '../sections/onboarding/ProductTypesSection';
@@ -25,7 +24,6 @@ export class OnboardingWizard extends BasePage {
 	productTypes: ProductTypeSection;
 	business: BusinessSection;
 	themes: ThemeSection;
-	benefits: BenefitsSection;
 
 	constructor( page: Page ) {
 		super( page );
@@ -34,7 +32,6 @@ export class OnboardingWizard extends BasePage {
 		this.productTypes = new ProductTypeSection( page );
 		this.business = new BusinessSection( page );
 		this.themes = new ThemeSection( page );
-		this.benefits = new BenefitsSection( page );
 	}
 
 	async skipStoreSetup() {