+++ packages/admin-e2e-tests/src/pages/WcHomescreen.ts
@@ -30,7 +30,7 @@ export class WcHomescreen extends BasePage {
 		await page.waitForSelector(
 			'.woocommerce-task-card .woocommerce-task-list__item-title'
 		);
-		await waitForElementByText( 'p', 'Finish setup' );
+		await waitForElementByText( 'p', 'Get ready to start selling' );
 		const list = await this.page.$$eval(
 			'.woocommerce-task-card .woocommerce-task-list__item-title',
 			( items ) => items.map( ( item ) => item.textContent )
@@ -45,7 +45,7 @@ export class WcHomescreen extends BasePage {
 	}
 
 	async clickOnTaskList( taskTitle: string ) {
-		const item = await waitForElementByText( 'div', taskTitle );
+		const item = await waitForElementByText( 'span', taskTitle );
 
 		if ( ! item ) {
 			throw new Error(