+++ packages/components/src/table/index.js
@@ -152,6 +152,7 @@ class TableCard extends Component {
 			summary,
 			title,
 			totalRows,
+			rowKey,
 		} = this.props;
 		const { showCols } = this.state;
 		const allHeaders = this.props.headers;
@@ -166,7 +167,12 @@ class TableCard extends Component {
 		return (
 			<Card className={ classes }>
 				<CardHeader>
-					<Text variant="title.small" as="h2">
+					<Text
+						variant="title.small"
+						as="h2"
+						size="20"
+						lineHeight="28px"
+					>
 						{ title }
 					</Text>
 					<div className="woocommerce-table__actions">
@@ -238,6 +244,7 @@ class TableCard extends Component {
 							caption={ title }
 							query={ query }
 							onSort={ onSort || onQueryChange( 'sort' ) }
+							rowKey={ rowKey }
 						/>
 					) }
 				</CardBody>
@@ -350,6 +357,11 @@ TableCard.propTypes = {
 	 * The total number of rows (across all pages).
 	 */
 	totalRows: PropTypes.number.isRequired,
+	/**
+	 * The rowKey used for the key value on each row, this can be a string of the key or a function that returns the value.
+	 * This uses the index if not defined.
+	 */
+	rowKey: PropTypes.func,
 };
 
 TableCard.defaultProps = {