+++ packages/onboarding/src/index.js
@@ -1,0 +1,7 @@
+export * from './components/WCPayCard';
+export * from './components/RecommendedRibbon';
+export * from './components/SetupRequired';
+export * from './components/WCPayAcceptedMethods';
+export { default as WCPayLogo } from './images/wcpay-logo';
+export { WooPaymentGatewaySetup } from './components/WooPaymentGatewaySetup';
+export { WooPaymentGatewayConfigure } from './components/WooPaymentGatewayConfigure';