+++ src/Loader.php
@@ -425,12 +425,6 @@ class Loader {
 		);
 		wp_style_add_data( 'wc-components', 'rtl', 'replace' );
 
-		wp_register_style(
-			'wc-components-ie',
-			self::get_url( 'components/ie', 'css' ),
-			array(),
-			$css_file_version
-		);
 		wp_style_add_data( 'wc-components-ie', 'rtl', 'replace' );
 
 		wp_register_style(
@@ -469,9 +463,9 @@ class Loader {
 		);
 
 		wp_register_style(
-			'wc-admin-ie',
-			self::get_url( "ie/style{$rtl}", 'css' ),
-			array( WC_ADMIN_APP ),
+			'wc-onboarding',
+			self::get_url( 'onboarding/style', 'css' ),
+			array(),
 			$css_file_version
 		);
 	}
@@ -697,6 +691,7 @@ class Loader {
 		wp_enqueue_script( WC_ADMIN_APP );
 		wp_enqueue_style( WC_ADMIN_APP );
 		wp_enqueue_style( 'wc-material-icons' );
+		wp_enqueue_style( 'wc-onboarding' );
 
 		// Use server-side detection to prevent unneccessary stylesheet loading in other browsers.
 		$user_agent = isset( $_SERVER['HTTP_USER_AGENT'] ) ? $_SERVER['HTTP_USER_AGENT'] : ''; // phpcs:ignore sanitization ok.