+++ tests/e2e/specs/activate-and-setup/complete-onboarding-wizard.test.js
@@ -1,9 +1,7 @@
 const {
 	testAdminOnboardingWizard,
-	testAdminNonUSRecommendedFeatures,
 	testSelectiveBundleWCPay,
 } = require( '@woocommerce/admin-e2e-tests' );
 
 testAdminOnboardingWizard();
-testAdminNonUSRecommendedFeatures();
 testSelectiveBundleWCPay();